feat: integrate email campaigns, job scheduler, Gmail API, GitHub API, and reminder email dispatch
- Add email campaign platform: organizations, campaigns, scheduled_emails schema (0003)
- Add job scheduler & workflow automation platform schema (0004)
- Integrate official Google Gmail API (GmailApiService + GmailApiController)
- Integrate official GitHub API (GitHubApiService + GitHubApiController)
- Implement job/workflow CRUD and execution endpoints (JobController + JobService)
- Implement campaign scheduling CRUD endpoints (CampaignController + CampaignService)
- Implement ReminderBackgroundWorker to dispatch due reminders via Gmail API
- Update OpenAPI contract to v1.4.0 with all new endpoint definitions
- Fix JsonMapConverter to use String binding for jsonb columns
- Fix FieldServiceControllerTest integer overflow assertions
- All 32 integration tests pass

+++ src/test/java/com/procureiq/springboot_app/FieldServiceControllerTest.java
@@ -148,8 +148,8 @@ public class FieldServiceControllerTest {
                 .content(objectMapper.writeValueAsString(assignRequest)))
                 .andExpect(status().isCreated())
                 .andExpect(jsonPath("$.status", is("success")))
-                .andExpect(jsonPath("$.data.serviceAppointmentId", is(appointmentId.intValue())))
-                .andExpect(jsonPath("$.data.serviceResourceId", is(resourceId.intValue())));
+                .andExpect(jsonPath("$.data.serviceAppointmentId", is(appointmentId)))
+                .andExpect(jsonPath("$.data.serviceResourceId", is(resourceId)));
 
         assertEquals(1, assignedResourceRepository.count());
     }
